+++ supervisor/registry.py
@@ -18,7 +18,7 @@ from config.settings import get_model
 from supervisor.types import AgentRunner, AgentTask, AgentResult
 from supervisor.llm_utils import _call_llm, _format_dep_context
 from supervisor.planner import _run_planner
-from supervisor.runners import _run_researcher, _run_coder, _run_critic, _run_summarizer, _run_tool_caller
+from supervisor.pipeline.runners import _run_researcher, _run_coder, _run_critic, _run_summarizer, _run_tool_caller
 
 log = logging.getLogger("goat2.supervisor")